backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
PyJWT==2.8.0
passlib==1.7.4
python-multipart==0.0.6
httpx==0.25.2
langchain
langgraph
openai
